backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
flask-jwt-extended==4.7.1
mysql-connector-python==9.2.0
bcrypt==4.3.0
python-dotenv==1.1.0

Flask-SocketIO==5.5.0
simple-websocket==1.1.0
matplotlib==3.10.1
